Remove requirement for application side double-buffering
At one time it was thought that application side double-buffering was
required for fast tracks to work reliably.  It now appears that this
was overly conservative, and that fast tracks work just fine with
single buffers on application side.  This is because the moment after
FastMixer pulls the data from the fast track, it synchronously signals
the application's callback thread to wake up, which will then run the
buffer completion callback handler to re-fill the buffer.  So disabling
this check now.

Bug: 9276275
Change-Id: I2c74a8c6d752140ccc7d285c5c5c1f7fcd792650