Check that the ABQ mime type matches the expectations
The code in audioPlayer_getAndroidObjectTypeForSourceSink assumes
that an android buffer queue source combined with an output mix
sink is an mpegts source, while it it assumes that it is an adts
source if it is combined with a simple buffer queue sink. This
assumption never was verified before, making the set up actually
seem to succeed even though the data is interpreted as the wrong
format.

Change-Id: I7e031ec8139fa2d770737935eb9acd3506d3fa3f