Build libOpenSLES for Brillo.
Currently, building libOpenSLES on Brillo requires a bunch of repos that
are not part of it. The dependency on libstagefright_http_support adds
these extra repos.

Break these dependencies by not calling
CreateHTTPServiceInCurrentJavaContext in Brillo. This does not change
the behavior in Brillo as CreateHTTPServiceInCurrentJavaContext returns
NULL if there is no Java VM.

BUG: 24578244
TEST=slesTest_sawtoothBufferQueue works on the dragonboard

Change-Id: I46479e3d223aedc7286eee2329baa507a812c386